fix: correct payload size parsing from integer to float
Changed the parsing function for MAX_PAYLOAD_SIZE to use parseFloat instead of parseInt. This allows the system to correctly handle payload sizes that are not whole numbers.

+++ lib/env.mjs.ts
@@ -7,7 +7,7 @@ export const env = createEnv({
 		MAX_PAYLOAD_SIZE: z
 			.string()
 			.optional()
-			.transform((value) => (value === "" || value === undefined ? 2 : Number.parseInt(value)))
+			.transform((value) => (value === "" || value === undefined ? 2 : Number.parseFloat(value)))
 			.refine((value) => value > 0, {
 				message: "MAX_PAYLOAD_SIZE must be greater than 0"
 			})
